requirements.txt
fastapi
uvicorn
python-telegram-bot[job-queue]
